+++ pkg/clusterfile/util.go
@@ -22,7 +22,7 @@ import (
 	"github.com/labring/sealos/pkg/constants"
 	yaml2 "github.com/labring/sealos/pkg/utils/yaml"
 
-	"github.com/labring/sealos/pkg/runtime"
+	runtime "github.com/labring/sealos/pkg/runtime/kubernetes"
 
 	k8sV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/yaml"
